+++ examples/arm64-job-definition/requirements.txt
@@ -1,2 +1,3 @@
 pulumi>=3.0.0,<4.0.0
-pulumi-aws>=6.0.2,<7.0.0
+pulumi-aws>=5.0.0,<6.0.0
+pulumi-awsx>=1.0.0,<2.0.0
